+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/commands/modify/shared/SelectionResolver.java
@@ -1,8 +1,8 @@
 package de.relluem94.minecraft.server.spigot.essentials.commands.modify.shared;
 
 import de.relluem94.minecraft.server.spigot.essentials.RelluEssentials;
-import de.relluem94.minecraft.server.spigot.essentials.constants.MessageKey;
-import de.relluem94.minecraft.server.spigot.essentials.helpers.objects.Selection;
+import de.relluem94.minecraft.server.spigot.essentials.enums.MessageKey;
+import de.relluem94.minecraft.server.spigot.essentials.model.Selection;
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.Nullable;
