+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/model/recipe/Shaped.java
@@ -1,6 +1,7 @@
-package de.relluem94.minecraft.server.spigot.essentials.helpers.recipe;
+package de.relluem94.minecraft.server.spigot.essentials.model.recipe;
+
+import org.bukkit.Material;
 
 import java.util.Map;
-import org.bukkit.Material;
 
 public record Shaped(String[] rows, Map<Character, Material> ingredients) {}
