+++ src/test/java/de/relluem94/minecraft/server/spigot/essentials/helpers/db/mapper/SettingMapperTest.java
@@ -1,7 +1,7 @@
 package de.relluem94.minecraft.server.spigot.essentials.helpers.db.mapper;
 
 import de.relluem94.minecraft.server.spigot.essentials.constants.Constants;
-import de.relluem94.minecraft.server.spigot.essentials.helpers.pojo.SettingEntry;
+import de.relluem94.minecraft.server.spigot.essentials.model.pojo.SettingEntry;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
